small tweaks and reorganization
--adicionado comentarios
--reorganizado algumas partes de código
--adicionado uma chamada ao websocket no codigo principal
