+++ external_modules/nano/nanoWebSocket.js
@@ -9,11 +9,11 @@ const ws = new ReconnectingWebSocket(('ws://[::1]:57000'), [], {
 	maxRetries: 100000,
 	maxReconnectionDelay: 2000,
 	minReconnectionDelay: 10 // if not set, initial connection will take a few seconds by default
-});
+})
 
 // As soon as we connect, subscribe to block confirmations
 ws.onopen = () => {
-	console.log("sdgasdgasdg")
+	console.log('websocket listening')
 	const confirmation_subscription = 
 	{
 		"action": "subscribe",
@@ -22,24 +22,20 @@ ws.onopen = () => {
 		{
 			//coloca todas as contas locais dentro do filtro para serem observadas se realizaram uma operacao
 			"all_local_accounts": true,
-			"accounts": [
-			//array de contas que vao ser observadas alem das contas locais
-			]
-		}
+			//"accounts": [array de contas que vao ser observadas alem das contas locais]			
+			}
 	}
-	ws.send(JSON.stringify(confirmation_subscription));
-	
-};
+	ws.send(JSON.stringify(confirmation_subscription));	
+}
 ws.onerror = function(event) {
 	console.error("WebSocket error observed:", event);
-  };
+  }
 
 // The node sent us a message
 //TODO:tratar os dados e enviar para o servidor e banco de dados
 ws.onmessage = msg => {
 	console.log(msg.data);
-	data_json = JSON.parse(msg.data);
-	if (data_json.topic === "confirmation") {
-		console.log ('Confirmed', data_json.message.hash)
-	}
-};
+	data_json = JSON.parse(msg.data)
+	
+	console.log ('Confirmed', data_json.message.hash)
+	};
